+++ guillotina/api/login.py
@@ -11,8 +11,11 @@ from guillotina.events import UserRefreshToken
 from guillotina.interfaces import IApplication
 from guillotina.interfaces import IContainer
 from guillotina.response import HTTPUnauthorized
+from guillotina.response import HTTPPreconditionFailed
 from guillotina.utils import get_authenticated_user
-
+from guillotina.auth.recaptcha import RecaptchaValidator
+from guillotina.auth.utils import find_user
+from json.decoder import JSONDecodeError
 import jwt
 
 
@@ -84,3 +87,221 @@ class Refresh(Service):
         await notify(UserRefreshToken(user, jwt_token))
 
         return {"exp": data["exp"], "token": jwt_token}
+
+
+@configure.service(
+    context=IContainer,
+    method="POST",
+    permission="guillotina.Logout",
+    name="@logout",
+    summary="Logout application",
+    allow_access=True,
+)
+@configure.service(
+    context=IApplication,
+    method="POST",
+    permission="guillotina.Logout",
+    name="@logout",
+    summary="Logout application",
+    allow_access=True,
+)
+class Logout(Service):
+    async def __call__(self):
+        user = get_authenticated_user()
+        try:
+            await user.logout(self.request)
+        except AttributeError:
+            pass
+
+@configure.service(
+    context=IContainer,
+    name="@users/{user}/reset-password",
+    method="POST",
+    permission="guillotina.Public",
+    responses={
+        "200": {
+            "description": "Reset password",
+            # TODO: add response content schema here
+        }
+    },
+    summary="Reset password",
+    allow_access=True,
+)
+class ResetPasswordUsers(Service):
+    async def __call__(self):
+        user_id: str = self.request.matchdict["user"]
+        actual_user = get_authenticated_user()
+        if actual_user.id == user_id:
+            # Self setting password
+            # Payload : {
+            #   'old_password': 'secret',
+            #   'new_password': 'verysecret',
+            # }
+
+            data = await self.request.json()
+            try:
+                await actual_user.set_password(
+                    data.get('new_password', None),
+                    old_password=data.get('old_password', None)
+                )
+            except AttributeError:
+                raise HTTPNotAcceptable()
+        else:
+            # We validate with recaptcha
+            validator = RecaptchaValidator()
+            status = await validator.validate()
+            if status is False:
+                raise HTTPUnauthorized(content={"text": "Invalid validation"})
+
+            # We need to validate is a valid user
+            user = await find_user({'id': user_id})
+
+            if user is None:
+                raise HTTPUnauthorized(content={"text": "Invalid operation"})
+
+            # We need to validate is a valid user
+            try:
+                email = user.properties.get('email', user.email)
+            except AttributeError:
+                email = None
+            if email is None and '@' in user_id:
+                email = user_id
+
+            if email is None:
+                raise HTTPPreconditionFailed(
+                    content={
+                        "reason": "User without mail configured"
+                    }
+                )
+
+            # We need to generate a token and send to user email
+            validation = app_settings.get('validation_process', None)
+            if validation is not None:
+                redirect_url = self.request.query.get('redirect_url', None)
+                await validation.start(
+                    as_user=user_id,
+                    from_user=actual_user.id,
+                    email=email,
+                    task_description='Reset password',
+                    task_id='reset_password',
+                    context_description=self.context.title,
+                    redirect_url=redirect_url)
+            else:
+                raise HTTPNotAcceptable()
+
+
+@configure.service(
+    context=IContainer,
+    name="@validate/{token}",
+    method="POST",
+    permission="guillotina.Public",
+    responses={
+        "200": {
+            "description": "Validate operation",
+        }
+    },
+    summary="Validate operation",
+    allow_access=True,
+)
+class ValidateOperation(Service):
+    async def __call__(self):
+        validation = app_settings.get('validation_process', None)
+        if validation is not None:
+            try:
+                request_payload = await self.request.json()
+            except JSONDecodeError:
+                request_payload = None
+            payload = await validation.finish(
+                token=self.request.matchdict["token"],
+                payload=request_payload)
+            return payload
+        else:
+            raise HTTPNotAcceptable()
+
+
+@configure.service(
+    context=IContainer,
+    name="@users",
+    method="POST",
+    permission="guillotina.Public",
+    responses={
+        "200": {
+            "description": "Register a user",
+            # TODO: add response content schema here
+        }
+    },
+    summary="Register Users",
+    allow_access=True,
+)
+class RegisterUsers(Service):
+    async def __call__(self):
+        allowed = app_settings.get('allow_register', False)
+        if allowed is False:
+            raise HTTPUnauthorized(content={"text": "Not allowed registration"})
+
+        validator = RecaptchaValidator()
+        status = await validator.validate()
+        if status is False:
+            raise HTTPUnauthorized(content={"text": "Invalid validation"})
+
+        payload = await self.request.json()
+
+        user_id = payload.get('id', None)
+        user = await find_user({'id': user_id})
+        if user is not None:
+            raise HTTPUnauthorized(content={"text": "Invalid login"})
+
+        validation = app_settings.get('validation_process', None)
+        if validation is not None:
+            redirect_url = self.request.query.get('redirect_url', None)
+            username = payload.get('fullname', payload.get('id', ''))
+            task_description = f"Registering user {username}"
+            actual_user = get_authenticated_user()
+            await validation.start(
+                as_user=payload.get('id'),
+                from_user=actual_user.id,
+                task_description=task_description,
+                task_id='register_user',
+                email=payload.get('email'),
+                context_description=self.context.title,
+                redirect_url=redirect_url,
+                data=payload)
+        else:
+            raise HTTPNotAcceptable()
+
+
+@configure.service(
+    context=IContainer,
+    name="@info",
+    method="GET",
+    permission="guillotina.Public",
+    responses={
+        "200": {
+            "description": "Get public information about user access",
+            # TODO: add response content schema here
+        }
+    },
+    summary="Info Access",
+    allow_access=True,
+)
+class InfoAccess(Service):
+    async def __call__(self):
+        validator = RecaptchaValidator()
+        status = await validator.validate()
+        if status is False:
+            raise HTTPUnauthorized(content={"text": "Invalid validation"})
+
+        auth_providers = app_settings.get('auth_providers', {})
+        providers = []
+        if 'twitter' in auth_providers:
+            providers.append('twitter')
+        if 'facebook' in auth_providers:
+            providers.append('facebook')
+        if 'google' in auth_providers:
+            providers.append('google')
+
+        return {
+            'register': app_settings.get('allow_register', False),
+            'social': providers,
+            'title': self.context.title
+        }