+++ guillotina/auth/__init__.py
@@ -1,4 +1,5 @@
 from .groups import GroupsUtility  # noqa
+from .recaptcha import RecaptchaValidator  # noqa
 from .utils import authenticate_request  # noqa
 from .utils import authenticate_user  # noqa
 from .utils import find_user  # noqa