+++ guillotina/contrib/dbusers/__init__.py
@@ -7,6 +7,9 @@ _ = MessageFactory("guillotina.contrib.dbusers")
 
 app_settings = {
     "auth_user_identifiers": ["guillotina.contrib.dbusers.users.DBUserIdentifier"],
+    "validation_tasks": {
+        "register_user": "guillotina.contrib.dbusers.register_user"
+    }
 }